# docs/api/openapi.json
{
  "components": {
...
    "schemas": {
      "AbandonSessionRequest": {
        "properties": {
          "reason": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "type": "object"
      },
...
      "AuthenticatedSession": {
        "properties": {
          "assurance": {
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "sessionToken": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "AuthenticationRequest": {
        "properties": {
          "loginIdentifier": {
            "maxLength": 64,
            "minLength": 3,
            "pattern": "[A-Za-z0-9._-]+",
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "loginIdentifier",
          "password"
        ],
        "type": "object"
      },
...
      "ChangePasswordRequest": {
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ],
        "type": "object"
      },
...
      "CreateSessionRequest": {
        "properties": {
          "amountMinor": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "currency": {
            "pattern": "^[A-Z]{3}$",
            "type": "string"
          },
          "lineSummary": {
            "maxLength": 1000,
            "minLength": 0,
            "type": "string"
          }
        },
        "required": [
          "amountMinor",
          "currency",
          "lineSummary"
        ],
        "type": "object"
      },
      "CreatedSessionView": {
        "properties": {
          "alreadyCreated": {
            "type": "boolean"
          },
          "checkoutId": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "sessionToken": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "SessionView": {
        "properties": {
          "amount": {
            "type": "string"
          },
          "checkoutId": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "lineSummary": {
            "type": "string"
          },
          "merchantId": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "paymentIntentId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The public HTTP contract. Versioned in the path (ADR-0015); errors follow RFC 9457, catalogued in docs/architecture/ERROR_CONTRACT.md. Generated from the application on every build and compared against this committed copy, so a change to the contract cannot reach a client without being reviewed.",
    "title": "finapp platform API",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/authentications": {
      "post": {
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthenticationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedSession"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
    "/v1/checkout/sessions": {
      "post": {
        "operationId": "createCheckoutSession",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSessionView"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
    "/v1/checkout/sessions/{id}": {
      "get": {
        "operationId": "viewCheckoutSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
    "/v1/checkout/sessions/{id}/abandonment": {
      "post": {
        "operationId": "abandonCheckoutSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AbandonSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            },
            "description": "OK"
          }
        }
      }
    },
...
    "/v1/me/credential": {
      "post": {
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthenticatedSession"
                }
              }
            },
            "description": "Created"
          }
        }
      }
    },
...
    "/v1/sessions/{id}": {
      "delete": {
        "operationId": "revokeSession",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  }
}